Reorganize processors structure
