+++ nft_ingester/src/lib.rs
@@ -1,9 +1,7 @@
-pub mod accounts_processor;
 pub mod ack;
 pub mod api;
 pub mod backfiller;
 pub mod batch_mint;
-pub mod bubblegum_updates_processor;
 pub mod buffer;
 pub mod config;
 pub mod error;
@@ -13,21 +11,16 @@ pub mod gapfiller;
 pub mod index_syncronizer;
 pub mod init;
 pub mod inscription_raw_parsing;
-pub mod inscriptions_processor;
 pub mod json_worker;
 pub mod message_handler;
 pub mod message_parser;
-pub mod mpl_core_fee_indexing_processor;
-pub mod mpl_core_processor;
-pub mod mplx_updates_processor;
 pub mod plerkle;
 pub mod price_fetcher;
+pub mod processors;
 pub mod raydium_price_fetcher;
 pub mod redis_receiver;
 pub mod rocks_db;
 pub mod scheduler;
 pub mod sequence_consistent;
 pub mod tcp_receiver;
-pub mod token_updates_processor;
 pub mod transaction_ingester;
-pub mod transaction_processor;
