+++ nft_ingester/src/processors/account_based/mod.rs
@@ -1,0 +1,5 @@
+pub mod inscriptions_processor;
+pub mod mpl_core_fee_indexing_processor;
+pub mod mpl_core_processor;
+pub mod mplx_updates_processor;
+pub mod token_updates_processor;
