+++ nft_ingester/src/processors/transaction_based/mod.rs
@@ -1,0 +1,1 @@
+pub mod bubblegum_updates_processor;
